rm buildEntities
