+++ compose/ui/ui/src/uikitMain/kotlin/androidx/compose/ui/window/SkikoUIView.kt
@@ -106,6 +106,7 @@ internal class SkikoUIView(
 
     init {
         multipleTouchEnabled = true
+        translatesAutoresizingMaskIntoConstraints = false
 
         _metalLayer.also {
             // Workaround for KN compiler bug
